Make `getIdFromToken` internal to OneToOneCallComposite
Cleanup of this composite is pending decision about its future.
