+++ packages/react-composites/src/composites/OneToOneCall/utils/SDKUtils.ts
@@ -7,8 +7,8 @@ import {
   PhoneNumberKind,
   UnknownIdentifierKind
 } from '@azure/communication-common';
-
 import { AudioDeviceInfo, LocalVideoStream, VideoDeviceInfo } from '@azure/communication-calling';
+import { CommunicationUiErrorCode, CommunicationUiError } from '../../../types/CommunicationUiError';
 
 export const getACSId = (
   identifier: CommunicationUserKind | PhoneNumberKind | MicrosoftTeamsUserKind | UnknownIdentifierKind
@@ -46,4 +46,40 @@ export const isLocalScreenShareSupportedInBrowser = (): boolean => {
     !isMobileSession() &&
     (/chrome/i.test(navigator.userAgent.toLowerCase()) || /safari/i.test(navigator.userAgent.toLowerCase()))
   );
+};
+
+// FIXME: This is hack to grab the userId from the token.
+//        Drop this in favor of explicitly passing in the userID to the composite.
+export const getIdFromToken = (jwtToken: string): string => {
+  const claimName = 'skypeid';
+
+  const jwtTokenParts = jwtToken.split('.');
+  if (jwtTokenParts.length !== 3) {
+    throw new CommunicationUiError({
+      message: 'Invalid jwt token',
+      code: CommunicationUiErrorCode.CONFIGURATION_ERROR
+    });
+  }
+
+  let base64DecodedClaims;
+  let base64DecodedClaimsAsJson;
+  try {
+    base64DecodedClaims = atob(jwtTokenParts[1]);
+    base64DecodedClaimsAsJson = JSON.parse(base64DecodedClaims);
+  } catch (error) {
+    throw new CommunicationUiError({
+      message: 'Invalid access token',
+      code: CommunicationUiErrorCode.CONFIGURATION_ERROR,
+      error: error
+    });
+  }
+
+  if (Object.prototype.hasOwnProperty.call(base64DecodedClaimsAsJson, claimName)) {
+    return `8:${base64DecodedClaimsAsJson[claimName]}`;
+  }
+
+  throw new CommunicationUiError({
+    message: 'Invalid access token',
+    code: CommunicationUiErrorCode.CONFIGURATION_ERROR
+  });
 };
